refactor: fix dirname of `ParseErrors`
