+++ src/plugins/editor-preview-asyncapi/components/ParseErrors/_parser-errors.scss
