+++ src/plugins/editor-preview-asyncapi/components/_all.scss
@@ -1,2 +1,2 @@
 @import './EditorPreviewAsyncAPI/editor-preview-async-api';
-@import './ParseErrrors/parser-errors';
+@import './ParseErrors/parser-errors';
